refactor(commandPalette): ♻️ make compostables smaller and more focused
