+++ skin.json
@@ -321,7 +321,14 @@
 				"resources/skins.citizen.commandPalette/components/CommandPaletteFooter.vue",
 				"resources/skins.citizen.commandPalette/components/CommandPaletteHeader.vue",
 				"resources/skins.citizen.commandPalette/composables/useKeyboardNavigation.js",
+				"resources/skins.citizen.commandPalette/composables/useItemNavigation.js",
+				"resources/skins.citizen.commandPalette/composables/useActionNavigation.js",
+				"resources/skins.citizen.commandPalette/composables/useScrollBehavior.js",
+				"resources/skins.citizen.commandPalette/composables/useKeyHandlers.js",
 				"resources/skins.citizen.commandPalette/composables/useSearch.js",
+				"resources/skins.citizen.commandPalette/composables/useSearchHistory.js",
+				"resources/skins.citizen.commandPalette/composables/useSearchNavigation.js",
+				"resources/skins.citizen.commandPalette/composables/useSearchResultTransformer.js",
 				{
 					"name": "resources/skins.citizen.commandPalette/icons.json",
 					"callback": "MediaWiki\\ResourceLoader\\CodexModule::getIcons",
